# apps/studio/frontend/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Junjo AI Studio",
    "description": "The Junjo AI Studio API: observability, execution evidence, and evaluation control.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.86.0"
  },
  "paths": {
...
    "/api/v1/auth-test": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Return the signed-in user's email.",
        "operationId": "auth_test",
        "responses": {
          "200": {
            "description": "The signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTestResponse"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sign-in": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with an email address and a password.",
        "operationId": "sign_in",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user is signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "The credentials are not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sign-out": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign out. This ends the session that made the request and no other.",
        "operationId": "sign_out",
        "responses": {
          "200": {
            "description": "The session is ended.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List every user, oldest first.",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Every user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create a user. Users are equal: any signed-in user may create, list, and\ndelete users.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The user was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The email address is already in use.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/create-first-user": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create the first user and sign them in.",
        "description": "Public so a fresh deployment can bootstrap. It is refused once any user\nexists.",
        "operationId": "create_first_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The first user was created and signed in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "A user already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{user_id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Delete one user. Their sessions and developer access tokens stop working\non the next request.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The user's identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "No signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthTestResponse": {
        "type": "object",
        "description": "The signed-in user.",
        "required": [
          "user_email"
        ],
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
      "CreateUserRequest": {
        "type": "object",
        "description": "The request to create a user.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable machine-readable code.",
            "examples": [
              "not_found"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable explanation."
          }
        }
      },
...
      "SignInRequest": {
        "type": "object",
        "description": "The request to sign in.",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "UserRead": {
        "type": "object",
        "description": "One Studio user. The password hash is never part of this type.",
        "required": [
          "id",
          "email",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier.",
            "examples": [
              "usr_2k4h6j8m9n0p1q2r"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address.",
            "examples": [
              "alice@example.com"
            ]
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the user account is active.",
            "examples": [
              true
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the user was created (UTC).",
            "examples": [
              "2025-01-15T10:30:00Z"
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the user was last updated (UTC).",
            "examples": [
              "2025-01-15T10:30:00Z"
            ]
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "description": "The response after a successful user operation.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "EvaluationControlToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A separately scoped Studio evaluation-control token. Studio ingestion API keys are not accepted."
      }
    }
  }
}